# services/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Life Platform API
  # 0.5.0 (P1-032): relaxes two request bodies, both loosening -- an
  # existing client keeps working. sign-in's familyId became optional
  # (omit it for a bootstrap session), and createFamily's ownerParentId
  # became optional (the session decides the owner; sending a mismatched
  # value is still refused). 0.4.0 (P1-031) added the five identity
  # operations. This is the REST API surface's own version, deliberately
  # independent of contracts/registry.yaml's contract_pack_version --
  # separate versioned surfaces per
  # docs/architecture/versioning-and-compatibility.md.
  version: 0.5.0
  description: >
    Canonical external REST contract for the Life Platform API, per
    docs/architecture/api-contracts.md. P0-005 established the reusable
    envelope/pagination/idempotency components and /health. P0-009 adds the
    Phase 1 contract pack: Family/ParentMembership/ChildProfile,
    TaskTemplate/TaskAssignment/TaskCompletion, VerificationResult,
    MediaEvidence, RewardLedgerEntry and Reward, with a representative (not
    exhaustive) set of paths -- enough for Backend/Parent Web/Child Web/QA
    to start building against a frozen shape per
    tasks/packets/P0-009-phase1-contract-pack.md. Full CRUD surfaces land
    with each entity's own P1 implementation task.
servers:
  - url: /api/v1
    description: Versioned API root. Breaking changes get a new version per
...
paths:
...
  /auth/sign-up:
    post:
      operationId: signUp
      summary: >-
        Create an adult account. Starts PENDING_VERIFICATION -- consent
        must be accepted before sign-in succeeds
        (docs/product/family-lifecycle.md). Issues no session. Failure is
        deliberately undifferentiated: a duplicate email reports the same
        error as any other problem, because confirming an address is
        registered is an account-enumeration oracle.
      tags: [identity]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 10
              additionalProperties: false
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                type: object
                required: [accountId, parentId, status]
                properties:
                  accountId:
                    type: string
                    format: uuid
                  parentId:
                    type: string
                    format: uuid
                  status:
                    type: string
        "400":
          $ref: "#/components/responses/Error"
        "5XX":
          $ref: "#/components/responses/Error"
  /auth/consent:
    post:
      operationId: acceptConsent
      summary: >-
        Record that required consent was accepted, moving the account to
        ACTIVE. This endpoint records only *that* consent happened -- what
        is consented to, under which rule, is a legal question owned by
        P1-034, and per docs/security/legal-ru.md nothing here supports a
        compliance claim.
      tags: [identity]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accountId]
              properties:
                accountId:
                  type: string
                  format: uuid
              additionalProperties: false
      responses:
        "200":
          description: Consent recorded; account is ACTIVE.
          content:
            application/json:
              schema:
                type: object
                required: [accountId, status]
                properties:
                  accountId:
                    type: string
                    format: uuid
                  status:
                    type: string
        "5XX":
          $ref: "#/components/responses/Error"
  /auth/sign-in:
    post:
      operationId: signIn
      summary: >-
        Exchange credentials for a session id. When familyId is given the
        session is scoped to a family the account is an ACTIVE member of
        -- authenticating proves who you are, not what you may act on.
        When it is omitted the result is a bootstrap session, which may
        only create a family (ADR-0006 constraint 3, DISC-P1-031-1).
        The returned sessionId is the opaque bearer value; it maps to a
        revocable record, not a signed token (ADR-0006 D2). Wrong
        password, unknown email, suspended account and non-membership all
        return the same failure. Throttled per email+client address.
      tags: [identity]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                familyId:
                  type: string
                  format: uuid
                  description: >-
                    Omit to receive a *bootstrap* session: an authenticated
                    parent who belongs to no family yet and may only create
                    one. Required for anything else -- see IssuedSession.
              additionalProperties: false
      responses:
        "200":
          description: Session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedSession"
        "401":
          $ref: "#/components/responses/Error"
        "5XX":
          $ref: "#/components/responses/Error"
  /auth/sign-out:
    post:
      operationId: signOut
      summary: Revoke the caller own session. Idempotent.
      tags: [identity]
      responses:
        "204":
          description: Session revoked.
        "5XX":
          $ref: "#/components/responses/Error"
  /auth/child-sessions:
    post:
      operationId: provisionChildSession
      summary: >-
        A parent provisions a session for their child device. Child
        access is never self-service: a ChildProfile carries no
        credentials by contract, so there is nothing a child could
        present (ADR-0006 D3). Requires a parent session, and the child
        must belong to that session family.
      tags: [identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childId]
              properties:
                childId:
                  type: string
                  format: uuid
              additionalProperties: false
      responses:
        "200":
          description: Child session issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssuedSession"
        "401":
          $ref: "#/components/responses/Error"
        "5XX":
          $ref: "#/components/responses/Error"
  /families:
    post:
      operationId: createFamily
      summary: Register a new family. Creates it in PENDING_INVITE status.
      tags: [family]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: >-
                ownerParentId is optional and vestigial: the owner is the
                authenticated actor. It predates sessions (P0-009), when a
                request had no other way to say who was acting. Sending it
                is allowed but it must match the session, so an older
                client cannot create a family owned by someone else.
              properties:
                ownerParentId:
                  type: string
              additionalProperties: false
      responses:
        "201":
          description: Family created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Family"
        "5XX":
          $ref: "#/components/responses/Error"
  /families/{familyId}:
    get:
      operationId: getFamily
      summary: >-
        Fetch a family. Authorization -- "Family is the security boundary
        for child data" (docs/product/actors-and-permissions.md): callable
        only by an ACTIVE member of this family or a scoped admin.
      tags: [family]
      parameters:
        - $ref: "#/components/parameters/FamilyIdPath"
      responses:
        "200":
          description: The family.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Family"
        "404":
          $ref: "#/components/responses/Error"
        "5XX":
          $ref: "#/components/responses/Error"
  /families/{familyId}/children:
    post:
      operationId: addChildToFamily
      summary: >-
        Add a child profile to a family. Requires the CHILD_POLICY
        capability (docs/product/family-lifecycle.md).
      tags: [family]
      parameters:
        - $ref: "#/components/parameters/FamilyIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [displayName, birthYear]
              properties:
                displayName:
                  type: string
                birthYear:
                  type: integer
              additionalProperties: false
      responses:
        "201":
          description: Child profile created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChildProfile"
        "5XX":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    FamilyIdPath:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    IssuedSession:
      type: object
      required: [sessionId, expiresAt, role]
      description: >-
        The opaque bearer value plus its lifetime. sessionId maps to a
        revocable server-side record (ADR-0006 D2) rather than being a
        self-contained signed token -- that is what makes
        docs/product/family-lifecycle.md revocation promise enforceable.
      properties:
        sessionId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        role:
          type: string
          enum: [parent, child]
        familyId:
          type: string
          format: uuid
          description: >-
            Absent on a bootstrap session. A client uses its absence to
            know it must create a family before anything else.
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code, e.g. "NOT_FOUND".
            message:
              type: string
              description: Human-readable, safe to display; never a raw
                driver/provider error string.
            details:
              type: object
              additionalProperties: true
              description: Optional structured context. Never PII, never a
                stack trace, never a provider payload.
            correlation_id:
              type: string
              description: Matches the request's correlation id, per
                docs/engineering/observability.md ("Correlation").

    # Reusable cursor-paginated list wrapper. Concrete list responses
    # compose this with an `items` schema of their own type.
...
    ParentCapability:
      type: string
      enum: [CHILD_POLICY, MONEY_REWARDS, SOCIAL_PERMISSIONS, CHAT_VISIBILITY, ACCOUNT_DELETION]
    ParentMembership:
      type: object
      required: [parentId, familyId, status, isFamilyOwner, capabilities, invitedAt]
      properties:
        parentId:
          type: string
        familyId:
          type: string
          format: uuid
        status:
          type: string
          enum: [INVITED, ACTIVE, REVOKED]
        isFamilyOwner:
          type: boolean
        capabilities:
          type: array
          items:
            $ref: "#/components/schemas/ParentCapability"
        invitedAt:
          type: string
          format: date-time
        activatedAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
    ChildProfile:
      type: object
      required: [childId, familyId, displayName, birthYear]
      properties:
        childId:
          type: string
          format: uuid
        familyId:
          type: string
          format: uuid
        displayName:
          type: string
        birthYear:
          type: integer
        avatarId:
          type: string
    Family:
      type: object
      required: [familyId, status, version, createdAt, parents, children]
      properties:
        familyId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING_INVITE, ACTIVE, SUSPENDED, ARCHIVED]
        version:
          type: integer
          minimum: 1
          description: >-
            Optimistic-concurrency token per
            docs/architecture/concurrency-and-conflicts.md. Submit the
            version you read; a stale version is rejected with an
            explicit conflict, never silently overwritten.
        createdAt:
          type: string
          format: date-time
        parents:
          type: array
          items:
            $ref: "#/components/schemas/ParentMembership"
        children:
          type: array
          items:
            $ref: "#/components/schemas/ChildProfile"
...
  responses:
    Error:
      description: Stable error envelope; never a raw database/provider error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []